Moved PlayersTable and Punishment classes to a folder
