+++ app/Livewire/Punishments/ShowPunishments.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Livewire;
+namespace App\Livewire\Punishments;
 
 use App\Helpers\TimeUtils;
 use App\Models\Player\Player;
